+++ src/Services/SuperListOrderService.java
@@ -7,6 +7,6 @@ import java.util.List;
 public interface SuperListOrderService {
     List<ListOrder> findAllListOrder();
     void addListOrder(ListOrder listOrder);
-    void update(int idOrder, double price, double grandTotal);
-    ListOrder getListOrderById(int id);
+    void update(Long idOrder, double price, double grandTotal);
+    ListOrder getListOrderById(long id);
 }
